make runCleanCompile.sh take cabal-dev sandbox as a parameter

+++ build-scripts/runCleanCompile.sh
@@ -1,14 +1,19 @@
 #!/bin/bash -e
 
+if [ -z "$1" ]; then
+  echo "Usage: $0 <cabal-dev sandbox directory path>"
+  exit
+fi
+
 rm -f kontrakcja-test.tix
-cabal update
-cabal clean
-cabal install --only-dependencies
-cabal configure -ftest-coverage
+cabal-dev --sandbox="$1" update
+cabal-dev --sandbox="$1" clean
+cabal-dev --sandbox="$1" install --only-dependencies
+cabal-dev --sandbox="$1" configure -ftest-coverage
 
 if [ "$TEAMCITY_VERSION" = "" ]; then
-  cabal build
+  cabal-dev --sandbox="$1" build
 else
-  cabal build 2>&1 | runghc build-scripts/Teamcity.hs ghc
+  cabal-dev --sandbox="$1" build 2>&1 | runghc build-scripts/Teamcity.hs ghc
   exit "${PIPESTATUS[0]}"
 fi
